renaming the unit test scripts
git-svn-id: http://svn.php.net/repository/pear/packages/Math_Stats/trunk@129926 c90b9560-bf6c-de11-be94-00142212c4b1
